Major rework of fourier transforms, power spectra and uetcs. Tried to seperate out functions such as projectors and fourier transforms to reduce code repetition. Futhermore it should be easier now to add vector uetcs.
We now have a seperate struct fft_fields that contain momentum space
fields and the fft_fields.in and fft_fields.out arrays required to
perform the ffts. This also contains the plan required to perform an
fft between the fft_fields.in and fft_fields.out fields.

Now ffts are performed seperately to power spectras.

The ffts themselves live in fft_func.c.

ALl projectors now live in projectors.c.

The respective power spectras live in scalarps.c, tensorps.c and
vectorps.c.

The uetcs still live in uetc.c.

Renamed lots of function names and outputs to try and represent what they do/are better.

Now we don't do the etc during uetc calculations, so we write out the
shear stress etc within using tensorps.c.  We now store the fourier
transform of the shear stress for the reference time uetc, so we don't
have to recompute this each time we do a uetc.  The plan is to perform
a fourier transform, then compute the etc, and then if required
compute the uetc, but by keeping the fourier transform outside the
power etc/uetc function, we can then reuse the transformed field for
the uetc.

Still to do:

- Double check that the uetc stills gives the same answer (when
  multiplied by factor of 32*pi).  It would appear that the real
  component does but the test on a small lattice I performed is
  inconclusive as the imaginary part was so small that it could have
  just been numerical noise.

- Rework the information that is printed out during the
FFT stage so it makes sense with our new structure.

- Print out timing information in a way that makes sense with new
  structure.

- Improve and proof read any new documentation.

- Add in vector uetcs.

- Think about how to have multiple reference times.

- Decide whether we want to normalise outputs in a different way: e.g
  change scalar ps to be normalised the same as vector and tensor ps.
  Also think about if we want to write out hdot ps and not have the
  factor of 1/(32 pi).

- Think about file reorganisation, should all power spectra go into a single powerspectra.c file?

- Implement changes into initps.c (maybe leave this to Pierre?)

- Should we have a seperate function, e.g etcs_and_uetcs(...), to
  carry out the specific set of function calls that occur within
  main.c to output the etcs and uetcs? It may look neater.

- If we do this we could seperate fft_func.c into fft_generic.c and
  fft_specific.c and put the above function and all of the fft_X,
  fft_e, fft_J and so on into fft_specific.c.

- Reduce code repetition by having a binning function?