Created uetc_vector and uetc_split_vector functions so that we may now compute vector uetcs. Implemented vector uetc for the velocity field.
Other changes:

We now normalise the scalar power spectrums the same as the vector and
tensor power spectra.

Always write out k value of bins rather than k*dx. Previously this was
not done for scalar ps and uetcs.

Changed naming convention for scalar power spectra to be in line with
vector and tensor.

Added in timing information for ffts. Still not particularly satisfied
with how this is done, may be a good idea to rethink timing for uetc
and power spectra and maybe for the whole code?