Changing fft data distribution so that cores responsible for each slab in x are equally spaced across the processor grid when the total number of cores is larger than Lx.
This should hopefully fix slow downs that occur when cores are all trying to send data to the first nodes on the the processor grid when the number of cores is much larger than Lx. Also should fix some memory bottlenecks that occur in this situation.

Still to do:
initps.c